# src/api/openapi.yaml
openapi: 3.1.0
info:
  title: Certificate Manager API
  description: API for managing SSL/TLS certificates
  version: 1.0.0
servers:
  - url: /api
    description: Main API endpoint
...
paths:
...
  /activity:
    get:
      summary: Get recent activities
      operationId: getActivities
      tags:
        - activity
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of activities to return
        - name: type
          in: query
          schema:
            type: string
            enum: [certificate, user, system]
          description: Filter activities by type
        - name: search
          in: query
          schema:
            type: string
          description: Search term to filter activities
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  activities:
                    type: array
                    items:
                      $ref: "#/components/schemas/Activity"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Add a new activity
      operationId: addActivity
      tags:
        - activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityCreate"
      responses:
        "201":
          description: Activity added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Clear all activities
      operationId: clearActivities
      tags:
        - activity
      responses:
        "200":
          description: Activities cleared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
        statusCode:
          type: integer
...
    Activity:
      type: object
      properties:
        id:
          type: string
          description: Unique activity identifier
        timestamp:
          type: string
          format: date-time
          description: When the activity occurred
        action:
          type: string
          description: Description of the activity
        type:
          type: string
          description: Type of activity (create, renew, delete, etc.)
        target:
          type: string
          description: Target of the action (certificate name, etc.)
        metadata:
          type: object
          description: Additional contextual information
    ActivityCreate:
      type: object
      required:
        - action
        - type
        - target
      properties:
        action:
          type: string
          description: Description of the activity
        type:
          type: string
          description: Type of activity (create, renew, delete, etc.)
        target:
          type: string
          description: Target of the action (certificate name, etc.)
        metadata:
          type: object
          description: Additional contextual information
...